app/requirements.txt
appnope==0.1.4
asttokens==2.4.1
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.0
executing==2.0.1
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kafka-python==2.0.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
wcwidth==0.2.13
